+++ src/main/java/com/TpFinal/view/publicacion/PublicacionAlquilerForm.java
@@ -18,6 +18,7 @@ import com.vaadin.event.ShortcutAction;
 import com.vaadin.icons.VaadinIcons;
 import com.vaadin.ui.*;
 import com.vaadin.ui.themes.ValoTheme;
+import org.vaadin.dialogs.ConfirmDialog;
 
 import java.time.LocalDate;
 import java.util.Collections;
@@ -169,16 +170,35 @@ public class PublicacionAlquilerForm extends FormLayout {
 
 
 
-    private void delete() {
-        service.delete(publicacionAlquiler);
-        addressbookView.updateList();
-        setVisible(false);
-        getAddressbookView().setComponentsVisible(true);
-        getAddressbookView().showSuccessNotification("Borrado");
-        getAddressbookView().enableGrid();
+
+        private void delete() {
+            ConfirmDialog.show(this.getUI(),"Eliminar Inmueble","Esta seguro que desea eliminar?, " +
+                            "esta accion no es reversible",
+                    "Entiendo y deseo eliminar","No eliminar",new ConfirmDialog.Listener() {
+                        @Override
+                        public void onClose(ConfirmDialog confirmDialog) {
+                            if (confirmDialog.isConfirmed()) {
+                                service.delete(publicacionAlquiler);
+                                addressbookView.updateList();
+                                setVisible(false);
+                                getAddressbookView().setComponentsVisible(true);
+                                getAddressbookView().showSuccessNotification("Borrado");
+                                getAddressbookView().enableGrid();
+
+                            }
+
+                        }
+
+                    });
 
 
-    }
+
+
+        }
+
+
+
+
 
    /* private void test() {
         nombre.setValue(DummyDataGenerator.randomFirstName());
